Move time slice CSV code to input submodule
Closes #288.
