+++ src/model.rs
@@ -5,7 +5,7 @@ use crate::commodity::Commodity;
 use crate::input::*;
 use crate::process::Process;
 use crate::region::Region;
-use crate::time_slice::{read_time_slice_info, TimeSliceInfo};
+use crate::time_slice::TimeSliceInfo;
 use anyhow::{ensure, Context, Result};
 use serde::Deserialize;
 use std::collections::HashMap;
